backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Messaging integrations
twilio==8.10.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pyjwt>=2.3.0
cryptography>=3.4.0
passlib>=1.7.4
tenacity>=8.0.0
schedule>=1.1.0
prometheus-client==0.19.0
